Split TestDependencyGraph into focused partial-class files
Fixes #10296

Split the 967-line TestDependencyGraph.cs into eight focused partial-class
files, each well under 300 lines:

- TestDependencyGraph.cs                              data + Build entry point
- TestDependencyGraph.EdgeResolution.cs               ResolveEdges
- TestDependencyGraph.CycleDetection.cs               FindCycles, FindCyclePathWithin
- TestDependencyGraph.SequentialSet.cs                ComputeSequentialSet, PropagateSequential
- TestDependencyGraph.ChunkBuilding.cs                BuildChunks, BuildChunkArrays
- TestDependencyGraph.ChunkCycleDetection.cs          HasChunkCycle, FindChunksOnCycle
- TestDependencyGraph.StronglyConnectedComponents.cs  shared iterative Tarjan pass
- TestDependencyGraph.Ordering.cs                     OrderTopologically, SelectIndices

Along the way, a few small cleanups that the split made obvious:

- FindCycles and FindChunksOnCycle carried two byte-identical copies of the
  iterative Tarjan traversal. They now share ForEachCyclicComponent, which
  invokes a callback once per strongly connected component that contains a
  cycle. The callback fires at exactly the same point in the traversal as the
  original inline code, so the order of reported cycles is unchanged.
- BuildChunks returned the exact same tuple from both sides of the
  chunk-cycle branch. The duplicate return is collapsed, and the comment
  explaining why the branch does not adjust what it returns now sits on the
  return it describes.
- HasChunkCycle now declares [NotNullWhen(true)] on its out parameters, which
  removes the two null-forgiving operators at its only call site.
- HasChunkCycle had two consecutive <summary> elements, which is invalid
  XML doc. The redundant first one is removed.

No behavioral change: all members stay internal, the InternalAPI surface is
untouched, and the unit and acceptance tests pass unchanged.

Copilot-Session: 2504b4aa-e01a-41e7-be86-60a0afb441bf
